Apply PIMPL idiom to vtkMaterialInterface classes.
After I added the DLL export for these classes, the Visual Studio
compiler started giving the warning "... needs to have dll-interface
to be used by clients of class" for each class member that was an
stl container.  Making the member into a pointer and moving the stl
code into the cxx file should make the warning go away.