Robustness improvements for vtkWrapClientServer.
I tried removing the BTX/ETX markers from the VTK header files, but
that caused the vtkWrapClientServer wrapping tool to choke on many of
the previously-blocked methods.  This patch makes vtkWrapClientServer
ignore the BTX/ETX makers, and makes it a bit smarter about deciding
which methods it should be able to wrap.