Fix for ICET compositing with non black backgrounds OpenGL2
Fixes transparent compositing when using backgrounds other than black.
Was a royal PITA to track down, seriously a PITA.

Basically background color was getting set on the FB and then
translucent geometry rendered on top of it. FB captured and
then composited with other FBs using a blend function assuming
alpha weighted colors (which is usual). But the background color
is not alpha weighted so two FBs, each with some background color
with translucent geometry on top can composite into something that
sums up to more than RGB 1,1,1. Producing wacky errors. Basically
the net effect was that the background color per pixel was getting
multiplied by the number of FBs that had translucent geometry on
that pixel.  If any FB had opaque geometry on top then it would hide
the error. If there was no translucent geometry the issue does not
happen. If the background color is black the issue does not happen.