Setting time step for vtkPVTrivialProducer properly for Catalyst.
When a Catalyst pipeline is initially set up the trivial producer
that is used to supply the vtkDataObject to the pipeline does
not have a time associated with it like the subsequent updates
to the trivial producer (done in UpdateProducers). This causes
issues with filters that are time dependent because the pipeline
doesn't think the producer provides any time information. This
update adds the time step while setting the vtkDataObject during
the CreateProducer call to avoid this issue.