Allow in situ particle paths to set terminal speed.
For particle paths the initial velocity field can be 0 and this
causes issues with the particle path filter since the initial
injected particles may be assumed to be stagnated and dropped
from integration at future time steps. This change allows
particle paths to set the terminal speed to an acceptable value.
Also moved the disableresetcache property from ParticlePath
to InSituParticlePath.