BUG: Fix repeated loading of state in Catalyst.
Once a modified state was sent to Catalyst from a ParaView Live session,
Catalyst ended up reloading the state on every iteration, when it really
only needed to load it when it changed i.e. a new state was sent over
by ParaView Live. This fixes that issue but not caching the previously
sent state.