Fix 'NonConvexPolygon' test with OpenGL2 backend.
This test was failing when we triangulated the input mest in
vtkPVGeometryFilter it was causing issues with the selection. In
reality, the need for explicitly triangulating the in the geometry
filter is no longer present for OpenGL2. The Mapper already takes care
of this and hence it's better to do it there. Hence, when using OpenGL2,
the 'Triangulate' option has no effect and the mapper will do whatever
is appropriate. This fixes the test failure as well. At the same time,
the bug #14308, which the "Triangulate" option was supposed to fix no
longer exists with OpenGL2 backend.