Getting the Catalyst examples testing properly.
This includes making the Catalyst examples build when BUILD_EXAMPLES
is enabled as well as doing testing when BUILD_TESTING is also
enabled. This will help test both Catalyst as well as the examples
to make sure that the examples aren't out of date. The labels for
the tests are PARAVIEW and CATALYST.