Fix widget order for 6 doubles.
pqDoubleVectorPropertyWidget was putting widget in a unintuitive order
when the property had 6 elements. Fixed that. How we place the widgets
so that the x-min, x-max are on 1st row, y-min, y-max on the next and so
on. The widget names are preserved so old tests should continue to work,
if any.