Break dependency cycles when unregistering proxies.
With addition of proxy list domain for "Glyph Type" property on the
Representation, we ended up with a cycle when unregistering proxies and
their dependencies. Simply creating a sphere (and a representaton) and
then deleting it was causing this infinite recursion. Keep state for
proxies being unregistered help break such cycles.