Improve Quick Launch behaviour.
The quicklaunch now returns the filters whom names start with the provided pattern
first. For instance, searching for "plot" shows "Plot Data" first, not "Extract Bag Plot".
Also, the quicklaunch pre-select the row of the first selectable item (if it exists),
not the first row (which can be not selectable because the filter is not acceptable
regarding the input).
Also, pressing CTRL+<SHIFT>+SPACE will auto-apply the selected source/filter from the
quick launch. This makes even more easier the creation of simple pipelines using the
quick launch dialog.