Adding a Interactive View Link Feature
This feature allow a user to check "Interactive View Link"
when creating a Camera Link.
This Interactive View Link takes the form of a widget inside the display
displaying the content of the linked view inside itself.

This is done only on the client side, and works with remote renderring.
It does not support mesa, and is not visible in collaborative mode or
tiling.