Fix a bug preventing paused insitu simulations from appling multiple changes
During InSitu visualization, there was a bug that caused the connection
between client and server to sever when the simulation was paused and more
than one parameter was changed before clicking "apply". The problem involved
multiple partially-updated state configurations being sent from the client
to the server when changes were applied. This fix tethers the shipping of
updated states to the "apply" gui command, ensuring that the updated state
is sent from client to server only once.