Tune FlatTreeView sizing
When running ParaView on HiDPI screens, where the IndentWidth may be
smaller than the font height, lines in the Pipeline browser
may get squashed together.

Fix this by defaulting to a preferredHeight for FlatTreeView items which
is the maximum between the font height and the IndentWidth, and only
adjusting upwards for items with a larger font.