Move the point gaussian representation into ParaView core.
Also add a new Python test of the representation that attempts to
exercise many of the features.  In addition, update pwf editor to
work better for the point gaussian representation, by adding X
range edit fields to the widget so that the user is not stuck
with X in [0, 1].  The change also ensures that user changes to
the piecewise function are propagated to the SMProperty on the
proxy, which was not the case before.