Fail gracefully when subsampling composite datasets.
The parallel implementation of vtkPExtractVOI and friends cannot
handle composite datasets. We work around this by running the
serial version of the filters when possible (that is, when sample
rate = [1, 1, 1]), but if subsampling is needed, the filters will
crash or hang. Detect this case and fail gracefully instead.

VTK bump disables the parallel version of the subfilters when
sample rate is [1, 1, 1].