BUG #15763: Fix mismatch in re-render state in client-server.
Avoid using camera's MTime to determine if the camera has changed. The
rendering pipeline is fraught with calls that change the MTime when the
camera may not have necessarily changed. That was causing the
client/server to mismatch when deciding if the selection cache was to be
cleared or not. Instead, we now simply use the mechanism already
provided on the vtkSMRenderViewProxy side to clear cache forcibly when
the camera sub-proxy is modified. This removes the dependence on
camera's Mtime, as well as ensures that the cache is used for
repeated selections when camera proxy isn't modified.