BUG #15867: Parallel projection wasn't restored from settings.
The problem was that the parallel projection mode was getting
overridden in vtkPVRenderView when initializing the interactor after
the property values had been loaded from settings. Hence the UI showed
correct value for parallel projection, however the Camera was still
doing perspective projection.

Since vtkPVRenderView changes parallel projection mode internally,
it makes sense to update the RenderView proxy XML to not directly set
parallel projection mode on the vtkCamera but instead pass it via vtkPVRenderView.
Thus vtkPVRenderView can cache it and restore it correctly when  interaction
mode is changed.