BUG #15843: Add support to test OpenGL version.
Added checks during initialization of a session to validate OpenGL
version support. If not adequate a dialog will be popped up warning
about the same. If the OpenGL version of server side alone is
inadequate, then remote rendering is disabled. If OpenGL version on
client is inadequate, then we pop up a dialog and abort.