animation: modify camera view up when motion starts verically
When the first motion of the camera is close to the camera's view up,
assume the camera's view up should always be tangent to the camera's
direction of motion.  This prevents strange view up modifications on
simple motions such as a vertical orbit where the view up can easily
line up with the vector from the camera's position to focal point.  Now
when the initial motion is within 45 degrees of the view up, the camera
will use the derivative of the motion as the view up during the
animation.