Allow (again) to choose point coordinates components as X Array Name.
Up until 4.4 the plot char views allowed to choose a point coordinates
component for the "X Array Name" - ie for the X axis of the plot.
This feature was non-intentionously removed.

This patch address the issue by adding the 3 point coordinates components
in the vtkSMChartSeriesListDomain (the underlying domain for the property "X Array Name").

To avoid code duplication I added a new method PopulateArrayComponents()
to generate the list of available array components in both Point/Cell
arrays and Point array coordinates.