Support reloading of files
This commit adds support to all reader to reload files. Both, reloading
already loaded files, or detecting new files in a file series is
supported.

To support reloading already loaded files in a reader-independent way,
we add support for `vtkSMProxy::RecreateVTKObjects()`. With help from
vtkSIProxy, this destroys and re-creates the VTK objects that the proxy
refers to and loads the proxy's state on it. If a reader supports
reloading in a more efficient way, then it can use the new
`<ReloadFiles>` proxy hint.

To make the RecreateVTKObjects() clean, this commit reorganizes the code
to create VTK objects in vtkSIProxy and subclasses. This should have
minimal impact on custom applications since it's not common to subclass
vtkSIProxy.

The reloading logic is encapsulated in `vtkSMReaderReloadHelper` which is
used by the new `pqReloadFilesReaction`. A new menu option in **Edit**
menu, as well as the context-menu in the pipeline browser can be used to
reload the active reader.