Check whether Catalyst has a live connection before doing work.
Currently if the Catalyst Python script has Live Vis enabled
then DoCoProcessing() is called every time whether or not
there is an actual connection for work to be done. This fixes
that by checking in RequestDataDescription to see if there is
a Live connection first to see if RequestDataDescription()
should be called. Because Live has the potential to change
output frequencies of the writer, we need to check those
frequencies instead of using the original frequencies.