BUG #15900: Fix parallel volume rendering.
The reported bug was manifesting because for the following:
+ Contour was requesting ghost cells from upstream.
+ vtkImageVolumeRepresentation would then reporte local extents (which
  included ghost cells) to the paralle redistribution code.
+ The parallel redistrbution code would fail to generate the KdTree
  since the extents with ghost cells overlap.

The fix was to make vtkImageVolumeRepresentation cache the extents etc.
in its RequestData. Now since vtkImageVolumeRepresentation (in its
superclass) requests EXACT_EXTENT from upstream, we know that ghost
cells won't be included in the image data passed to RequestData and
hence if we cache the extents then, won't see the above problem.

This fix did uncover BUG #15898 during testing, which has been reported.

The same fix is applicable to vtkImageSliceRepresentation too, which
uncovered another issue (reported as BUG #15899).