Fixing the processing of "ProxyList" hints.
"ProxyList" hints are used on proxies that can be part of a
vtkSMProxyListDomain to link certain properties on them to properties on
the parent proxy of the vtkSMProxyListDomain. This makes it possible for
things like using input data information to setup center for implicit
plane etc.

This was earlier done in vtkSMParaViewPipelineController which meant
these links only got setup when these proxies were created from the
application and not when loading state or undo/redo. That was a bug.
This fixes it by make vtkSMProxyListDomain itself process these hints
and set the links up.