BUG #15898: Fixes issue with volume rendering with ghost cells.
EXACT_EXTENT() in VTK pipeline doesn't necessarily provide exact extents
to a filter. To overcome that issue, vtkImageVolumeRepresentation must
explicitly handle the case when the extents it receives are large i.e
have ghost cells. vtkImageVolumeRepresentation now detects this explicitly
and removes the ghost cells before passing the data on to the volume
rendering code.