Making in situ particle paths more general and fix build issue.
The in situ particle path filter can have an arbitrary start
time now that is aware of the initial time step other than zero.
This can be used for restarted simulations or just ignoring
the first set of steps of a simulation. The example was updated
accordingly.
Buildwise, the Fortran Catalyst examples were disabled by default
since there is a build issue on some machines.