Fix the unserialization of double values from string.
atof() function is used to unserialize colors values from string.
The problem is that Qt use to set the locale according to the user's
environment settings. Thus, LC_NUMERIC could, for instance, be
set to "FR_fr.UT8" implying that the numerical separtor is ','.
This then breaks the unserialization mechanism which uses the
locale while the serialization does not.

This fix overrides the LC_NUMERIC environment variable to "C"
in the paraview application process at startup. Qt then uses
this locale instead of the system one.

function is locale dependent. It results that, on non US/EN, environment,
(eg. french) atof("0.456") returns 0 as the separator "," is expected
and not ".".
We changed atof() to a C++ unserialize from stream.