BUG #15882: Fixes performance of hover-points mode.
Hover-points-tooltips mode was slower than interactively selecting
points. The reason was that while we reused selection buffers for
latter, the former was re-rendering the selection buffers on each mouse
move. This was because the cache preserving mechanisms that the render
view has for preserving selection color buffers while interacting in an
interactive selection mode were not being employed for hover-points
mode. Fixed that by chaning the render view's interaction mode state so
as far as the render view goes, it's in the same mode as the interactive
point/cell selection and will use the same caching techniques.