BUG #15961: Fix blending issues for remote rendered images.
For the reported bug, the server-side background color gets blended into
the renderered image. When the image is delivered to the client and then
blended on to the image on the client side, we end up with the
background color getting blended in again. This was causing the
brightness issue reported. We address this bug by making server ranks
render with a black background always (when in modes where remote
rendered image is delivered to the client).