Set WriteSummaryFlag consistently on all ranks.
vtkXMLPVDWriter was not setting WriteSummaryFlag consistently on all
ranks. vtkXMLPDataWriter always had logic to only write data on root
node, so this logic was unnecessary here. When upcoming changes to
vtkXMLPDataWriter to do parallel reduction when WriteSummaryFlag is
ON, it becomes critical that WriteSummaryFlag is set consistently
everywhere. This address that.