Fix an issue at linking vtkPVCatalystPython when using MPI.
The class declares a method with a MPICommunicatorOpaqueComm object
which is declared in vtkMPI.h. When the wrapping layer code is
generated this adds a dependency on this header file and thus requires
to link the wrapped file with MPI libraries.

To avoid that, we exclude this method of the wrapping.