Revert default output precision type for Contour filter points
The default was changed to floats a while back to handle an apparent
regression when the input points type was integral (it happens), but
this wasn't great when the input point precision was double. This
commit changes the default value back to
vtkAlgorithm::DEFAULT_PRECISION, which sets the precision of the
output to the precision of the input.

The output precision type remains exposed for corner cases where the
the input precision type is an integral type. Such cases have come up
on the mailing list more than once, and the result is a truncation of
the contour coordinates to integral values, leading to cube-ish
contours.