BUG #15167: Show multiple blocks in spreadsheet view.
Spreadsheet representation now allows the user to choose multiple blocks
from a composite dataset to be shown in the view. Multiple blocks are
merged and hence will loose partial arrays, if any.

To still support picking the default value as the first non-empty leaf
node for Spreadsheet representation, added a mechanism to
vtkSMCompositeTreeDomain to select a default mode.