CameraLink should link parallel projection mode as well.
Since camera interaction parameters depend on the projection mode the
camera is in, when one links the camera interaction, one should also
link the projection mode since otherwise the link does not work.
This fixes that issue.

Updated the CameraLink test to test both perspective and parallel
projection when using camera links.