Avoid double dlclose().
Plugin unloading was not done properly. Consequently, it was possible to
call dlclose() on a plugin that was already unloaded during
finalization. This commits add guards against such double unloading
happening.

The fact that this was causing errors with Qt5 alone is coincidental.
One can indeed verify that the plugin library was unloaded before it got
to clean up logic in vtkPVPluginLoader even earlier by using
`LD_DEBUG=libs` environment variable.