BUG #15997: Add ability to save color information in PLY files.
The PLY writer now allows user to choose whether color information is to
be saved in the output file. When enabled, ParaView uses the data
sources coloring state in the active view i.e. the array to color with
and the LUT to use, to pass along to the writer so that it can save RGB
color values in the output dataset.