LagrangianParticleTracker Plugin
This plugin expose the vtkLagrangianParticleTracker,
a very flexible particle tracking filter in vtk,
as well as providing helper filter for easier use of the filter.
The helpers are :
 * A surface helpers, which add a certain number of values to the field data of it's input,
   to be used then as the surface in the tracker.
 * A seed helpers, in multiple version, which a certain number of values to the point data of it's input,
   to be used then as the seed input of the tracker.

Both these filters uses an integration model parameters to configure their panel, in order to be completelly dynamic.

This commits also contains lots of tests for the tracker and helpers.