Don't show search bar in pqProxyWidgetDialog unless useful.
If pqProxyWidgetDialog is asked to enableSearch, however there are no
advanced properties on the proxy, then it can be confusing for the user.
Hence, added a mechanism to pqProxyWidgetDialog to ask if there are any
advanced properties. Calling code can use that information to determine
if the search bar must be enabled.

Updated the writer dialog code (in pqSaveDataReaction) to not show the
search bar unless there are any advanced properties.

See BUG #16080.