Refactoring example vis state loading.
We now load the state file from the Qt resource system (avoid having to
look for those files). Data files looked up under directories are
supported by paraview/paraview-superbuild!108 and then state files
loaded if data is found. If not, a message is shown to notify the user
of the missing data files in their installation.