Add mechanism to override default data distribution mode.
Added experimental feature to support overriding default data
distribution mode selected by vtkPVRenderView from a representation.
This can be used, for example, if a representation wants all data to be
'collect'ed on root node.

THIS IS AN EXPERIMENTAL FEATURE AND ONLY ADDED TEMPORARILY AND MAY NOT BE
SUPPORTED LONG TERM.