BUG 15860: Restore to array-specific LUT when restoring default
Previously, when restoring the default transfer function, the one XML
default was always restored. Now, presets for specific array names are
supported in ParaView, so we reset to the preset if it is available.
In particular, this solves the problem of restoring the default
transfer function for the vtkBlockColors array when the defaults are
restored.