Use parallel methods to check file systems instead of serial methods.
For config information, plugins, etc. we are having each pvserver/pvbatch process
check the file system for information through VTK's SystemTools class. This is
very inefficient when going to moderate to high process counts. Changes are
to have process 0 check information and then broadcast that out. vtkPSystemTools
and vtkPDirectory are classes that have recently been added to VTK and that do that
very efficiently in parallel.