Disable propagation on (Start/End)InteractionEvent in vtkSMCameraLink.
vtkSMCameraLink propagated start/end interaction events through the
links. This was done a long time ago when we wanted to support "linked"
interaction undo/redo stacks -- something we no longer do.

Also, that's conceptually wrong. Observers listening to
StartInteractionEvent on a view may think a user is interactive with
multiple views at the same time is they are linked together via
vtkSMCameraLink.

Hence removing this logic to propagate these events. vtkSMCameraLink
separately manages whether to call InteractiveRender on StillRender on
the view and that part is unaffected.