Fix PlotMatrixViewArraySelection test.
The test playback when playing test through ctest or when manually playing the xml
resulted in different baselines. That was because the playback through
ctest playback caused the appliction to capture image even before the
"animation" in the plot matrix view was finished. This fixes that issue
by using the new "process_events" command in QtTesting to pause after
playting the test while processing events to let the animation finish.