Add user toggle for 2D log plot with negative data values.
This commit addresses issue # 0014895. Originally, when data with negative
values were plotted on a logarithmic scale, a warning would display until
the axis range was correctly modified to show the data. This commit adds a
persistent toggle that changes this behavior by rescaling the axis range to
appropriate values when logarithmic scale is selected.