Cleaning up vtkPVDataSetAttributesInformation.
Refactored how vtkPVDataSetAttributesInformation keeps track of the
arrays. We now keep them in a map which cleans up the code quite a bit.
It does mean two arrays with same name are treated as the same, which is
indeed the case for rest of the application.

The baselines changed due to this commit affecting the sorted order of
arrays. A bug in old code resulted in "Result_total" coming before
"Result_average" in alphabetical order which is clearly wrong. This
commit fixes that.