Ignore array information on empty datasets.
When combining information about blocks in a composite dataset, we now
no longer mark arrays as partial because they are missing on empty
datasets. Having an empty dataset or a NULL at a specific leaf node in a
composite dataset should now behave identically as far as partial array
indication goes.

Also added a test that tests this explicitly.