Fix errors in collaborative session.
If the `master` had an existing state when a slave client connects to
collaborative session, we would get debug messages from pqPipelineModel
warning us that it couldn't locate the server for the proxy added. This
happened because the initial collaboration state was being loaded before
the pqPipelineModel was told that a new session was created. Making
pqPipelineBrowserWidget use the signal fired before a session gets
initialized to let pqPipelineModel know of a new session.