vtkPVPluginTracker: only check for duplicate plugins by name
If a plugin is hard-coded to be loaded, it gets entered into the list of
loaded plugins as (name, "internal"). If the plugin is then loaded
explicitly by path, this would first find it (by name) and then not find
it because of the path. Instead, just ignore the path altogether; the
name should be sufficient.