Add global toggle to load all variables from a file by default.
This commit is in reference to issue # 0016027. Originally, the default values
selected for loading were decided by the file readers. This commit preserves
this functionality, but adds a global setting to load all values by default.
With this change, we do not make any distinction between data values and data
blocks/assemblies/materials; all are enabled by default.