full camera for cinema
This adds a new camera model where instead of
phi and theta tracks we have a single pose track that contains
the 3x3 orientation matrix for a camera that defines the direction along
with a record of initial placement at each timestep.

The placement defaults to stationary but can follow the position of a named
object in the pipeline.

Other changes prevent oversampling in theta, coarsen longitudinal sampling,
toward the pole, and allow higher resolution sphere samples.