BUG 15923: Gracefully handle empty Python Calculator expressions
Invoking a Python Calculator with an empty expression would produce
an error message about failure to parse a Python expression. While
an error message may seem reasonable, when the ParaView GUI program
is in "auto apply" mode, the error messsage pops up whenever a
Python Calculator is first added, which is not necessary.

The fix is simple - do not evaluate the expression if it is an
empty string.