updates to Cinema export
This change lets File->Export export specA (non composite) stores
like Catalyst Script Generator Plugin can.
It also allows creation of static (no camera motion) specB stores.
It also unifies the coprocessing output runtime path so that both
specA and specB no longer go through completely different code paths.
It also removed the behavior of supplying a default selection of
values for contours, clips and slices when the object is exported
but no values a chosen by the user.

One minor regression is that specA lost the code that chooses a
default camera position that is closest to the view shown when the
script was created.