15766: Enable locking down dock widgets
It is not uncommon to accidentally move a dock widget when interacting with
the UI. Getting the dock widget back to where it was can be a challenge.
Adds a setting to lock down the dock widgets so that they cannot be moved,
but can still be closed or reopened.