9356: Expose some of Connectivity filter's properties
The ExtractionMode and ColorRegions properties were hidden. Also,
added a ClosestPoint property for use when the closest point mode is
enabled.

A subset of extraction modes that are easily supported are exposed.
Modes that use point or cell seeds defined by selections could be
added in the future, but are not added here.