Prefer accessing toolbar buttons by name to by index
There are two ways for the buttons in a toolbar to be referenced by
tests, by the name of the button and by the position in the toolbar.
The position in the toolbar can change based on the width of the toolbar
since some items may be pushed into the 'expand toolbar' section.  And
that changes how they are accessed by index since QtTesting includes
visibility in the index access name for the button.  Accessing the
buttons by name is much safer since the window system will determine the
default size of the window (which affects the default size of the
toolbar and which buttons are shown).