Keep empty string annotation values from appearing in scalar bar
The ParaView UI represents annotation values and annotations as
strings. There was a bug in setting the full list of annotation values
when one or more of the string values read from the UI were the empty
string. The annotated value "" was considered acceptable, added to the
full list of annotations, and the result was that an undesirable
"extra" categorical color could appear in the color legend.

Fixed this by not adding an annotation when the its value is the empty
string. Note that this will prevent the empty string from being an
acceptable category in vtkStringArrays.