Ensure maximized/minimized widget is active.
When maximizing/minimizing a view frame during test playback, if the
view frame was not active already, the active view did not change. The
change happened correctly during user interaction due to the
mouse-click which doesn't get recorded in the test. Fixed by ensuring
that both these actions activate the frame, if it's not active already.

Addresses parts of issue #16824.