Use correct helper proxies group name when loading states
When loading state files, added mechanism to
change the proxy's registration group and name. We use this new
mechanism to change the groupname for helper proxies associated with a
particular proxy. That fixes #16862, #16863.

To support ability to change the groupname for helper proxies before we
registered them, we had to change the time when proxies are registered
when loading state. Previously they were registered as soon as they were
created. Instead, now, we register them in bulk after creation.  We
still preserve the registration order since that is the proxy dependency
order as well.

Another change is that state for timekeeper and animation proxies is
loaded after all other proxies have been created and registered. Since
timekeeper and some animation proxies are already registered, doing so
ensures that when their properties start getting modified, the referred
proxies are already registered.

Added a test (TestHelperProxySerialization) that tests the new
functionality.