Add ExtractTimeSteps filter
ExtractTimeSteps extracts the specified time steps from the input dataset.
The timesteps to be extracted are specified by their indices. If no time
step is specified, all of the input time steps are extracted. This filter
is useful when one wants to work with only a sub-set of the input time steps.

Addresses #16837