Fix #16052: Respect lock-data-range toggle.
ParaView was not respecting the `lock data range` toggle for transfer
functions in pqDisplayColorWidget when switching array to color with.
Fixed that. Added API args to
vtkSMPVRepresentationProxy::RescaleTransferFunctionToDataRange() that
control whether to force the LUT range or not and `pqDisplayColorWidget`
now uses that. Default behavior for `RescaleTransferFunctionToDataRange`
remains unchanged, so Python scripts etc. should continue to work as
earlier.