Remove alpha channel from plot image data
The alpha channel from the Python-generated image in the test had all
zero values. In the OpenGL 1 backend, this was apparently disregarded
as the Python-generated image was displayed just fine. In the OpenGL 2
backend, however, the alpha channel appears to be used. Because the
alpha was 0, the rendered image was 0. Fixed this bry dropping the 4th
alpha channel.