Fix 'File Dialog' issues with test playback on OsX
On OsX, when playing tests, any tests with file dialog exhibited
following issues:

1. With Qt 4, the test would segfault randomly
2. With Qt 5.7, the test would deadlock quite reliably.

In both cases, it seemed something to do with the what causes the
dialog's event loop to close since I could narrow it down
to `pqFileDialog::selectFile` call to `QDialog::accept`.

Instead of explicitly doing that, we now trigger a "click" on the OK
button. This seems to address the issue entirely. Is also consistent
with how other modal dialog behave during test playback.