Fix name for coordinates array in H5PartReader.
H5PartReader was naming the coordinates array same as the array chosen
to be the X array. That would cause problems with views like chart and
spreadsheet view where the array names are used for lookup.

Fixed that by ensuring that vtkH5PartReader uses a the standard name for
Points array i.e. `Points`.

Fix issue #16909.