hide the catalyst specific options when exporting from file menu
use better extension for top level cinema directory

.json is a horrible extension for a folder especially as it
confuses the cinema viewer. Instead, call it .cws for cinema
workspace as it contains one or more views making up a workspace.

The case of the disapearing camera samples - SOLVED

It was the old man from the candy shop scooby. He didn't provide
a sufficient minimum size for the row that help the phi,theta,roll
widgets. "And I would have gotten away with it if it wasn't for
those pesky kids!" Rooby dooby doo!