Changed default task in Descriptive and Contingency Statistics filters
The old default task was set to compute statistics over a subset of
the data. This yielded minima and maxima that aren't your conventional
minima and maxima of the entire data set. The worry was that folks may
misinterpret the minima and maxima for the subset as the minima and
maxima for the entire data set. Addressed this by setting the
default to "Detailed model of input data", which computes statistics
over the entire dataset.

Addresses #14886