Make Qt help optional for ParaView
The Qt Help module adds several dependent libraries, and is not
necessary for code reusing the ParaView libraries. This patch makes Qt's
help system optional, defaulting it to on in order to preserve existing
behavior, and exposed a PARAVIEW_USE_QTHELP option in a similar spirit
to other CMake flags.

In the Tomviz project this removed the links to the Help, CLucene, and
Sql Qt 5 libraries (along with other things they link to).