Empty commit before clang-format formatting
This commit is an empty commit that precedes the automatic application
of clang-format across the ParaView codebase.  This can be used to
rebase a topic across the transition to clang format.  First rebase your
topic onto this commit.  Then rewrite the topic, including this commit
with `git filter-branch --tree-filter` with
`Utilities/Scripts/clang-format.bash` to update the style in all commits
in the topic.  Finally rebase the updated topic, excluding the rewrite of
this commit onto the style transition commit.

See `git help filter-branch` for details.