Inform the camera of the IceT subsampling projection matrix.
Some classes (e.g. widgets) use vtkRenderer/vtkCoordinate to
convert display <--> world coordinates. However, on OpenGL2,
only the actual shader matrices were modified, and the camera state
was not touched, leading to rendering artifacts when interactive
subsampling was enabled in ParaView.

This uses the new ExplicitProjectionMatrix in vtkCamera to allow
the VTK coordinate converters to use the same transformation
pipeline as the shaders.