Enable quoting of variables to accommodate special characters
Array names may contain units inside parentheses, e.g., "Velocity
(m/s)". When they do, they are unavailable in the Calculator because
the parentheses and forward slash are interpreted as part of an
expression rather than as a variable. To address this, support using
quotation marks around such names so that they are interpreted as
variables.

Fixes #16923