Refactor recent files menu
This refactors recent files menu (`pqRecentFilesMenu`) in ParaView to make
it more extensible. Now, application can easily provides implementations
of an interface (`pqRecentlyUsedResourceLoaderInterface`) to add support
for additional types of items to add to this menu.  The new class
`pqStandardRecentlyUsedResourceLoaderImplementation` provides the
implementation for this interface for recent items tracked by ParaView
i.e. loading of data and state files.