Fix cell-centered glyphs
When glyphing with cell-data, the check is vtkPVGlyphFilter to ensures
that the scalars and vectors array chosen are both of the same
array-association was a bit too conservative. Thus is scalars was set to
point-data, vectors to cell-data, but we're only using vectors for
glyphing (as was the case with BUG #17081), then we'd get errors and the
glyph filter would give up. Fixed by adding checks to ensure that the
scalars (and vectors) and both needed before checking if they are of the
same association.

`Flow` test was also changed to handle the fact that the glyph filter no
longer marks input scalars as the active scalars on the output, which
affects how the filter's output is colored by default in ParaView.

Addresses #17081.