Invalid resource accessedin pqQVTKWidget.
During initialization, pqQVTKWidget was accessing an icon resource that
was not present in the pqCore library, instead only present in
pqComponents. Hence, when apps linked against pqCore alone e.g.
`pqCoreBasicApp`, it would raise an error attempting trying to paint on
a 0-sized image. Fixed that by creating a copy of the resource in
pqCore.